[Favorites] Adds categories to the favs/follows
