+++ src/lib/Scenes/Favorites/Components/Categories/index.tsx
@@ -2,32 +2,39 @@ import React from "react"
 import { FlatList, Text, View } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
 
-import SavedArtistRow from "../Artists/Components/SavedArtistRow"
+import SavedItemRow from "lib/Components/Lists/SavedItemRow"
 
-class Artists extends React.Component<RelayProps, null> {
-  renderRow(item) {
-    return (
-      <Text>
-        {item.id}
-      </Text>
-    )
-  }
-
+class Categories extends React.Component<RelayProps, null> {
   render() {
-    const rows: any[] = this.props.me.followed_genes.edges.map(edge => edge.node)
-    return <FlatList data={rows} keyExtractor={({ __id }) => __id} renderItem={this.renderRow.bind(this)} />
+    const rows: any[] = this.props.me.followed_genes.edges.map(edge => edge.node.gene)
+
+    return (
+      <FlatList
+        data={rows}
+        keyExtractor={({ __id }) => __id}
+        renderItem={data => <SavedItemRow square_image {...data.item} />}
+      />
+    )
   }
 }
 
 export default createFragmentContainer(
-  Artists,
+  Categories,
   graphql`
     fragment Categories_me on Me {
-      followed_genes {
+      followed_genes(first: 10) {
         edges {
           node {
-            id
-            __id
+            gene {
+              id
+              __id
+              name
+              href
+              id
+              image {
+                url
+              }
+            }
           }
         }
       }
